Remove 14.04 for Kraken and update 16.04 symlinks
