--- ceph-releases/kraken/ubuntu/14.04/daemon/2.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-PATH=/command:/usr/local/bin:/usr/local/sbin:/bin:/sbin:/usr/bin:/usr/sbin:/usr/X11R6/bin
-
-exec env - PATH=$PATH \
-runsvdir -P /etc/service 'log: ...........................................................................................................................................................................................................................................................................................................................................................................................................'
