--- ceph-releases/kraken/ubuntu/14.04/daemon/ceph.defaults
@@ -1,71 +1,0 @@
-# ceph.defaults
-# the format is key_path value
-# the key path will be prefixed by /ceph-config/$CLUSTER
-
-# auth
-/auth/cephx true
-/auth/cephx_require_signatures false
-/auth/cephx_cluster_require_signatures true
-/auth/cephx_service_require_signatures false
-
-# auth
-/global/max_open_files 131072
-/global/osd_pool_default_pg_num 128
-/global/osd_pool_default_pgp_num 128
-/global/osd_pool_default_size 3
-/global/osd_pool_default_min_size 1
-
-/global/mon_osd_full_ratio .95
-/global/mon_osd_nearfull_ratio .85
-
-#mon
-/mon/mon_osd_down_out_interval 600
-/mon/mon_osd_min_down_reporters 4
-/mon/mon_clock_drift_allowed .15
-/mon/mon_clock_drift_warn_backoff 30
-/mon/mon_osd_report_timeout 300
-
-#osd
-/osd/osd_journal_size 100
-
-# these 2 should be passed at runtime to the container.
-#/osd/cluster_network 198.100.128.0/19
-#/osd/public_network 198.100.128.0/19
-/osd/osd_mkfs_type xfs
-/osd/osd_mkfs_options_xfs "-f -i size=2048"
-/osd/osd_mkfs_options_btrfs "-f"
-/osd/osd_mon_heartbeat_interval 30
-
-#crush
-/osd/pool_default_crush_rule 0
-/osd/osd_crush_update_on_start true
-
-#backend
-/osd/osd_objectstore filestore
-
-#performance tuning
-/osd/filestore_merge_threshold 40
-/osd/filestore_split_multiple 8
-/osd/osd_op_threads 8
-/osd/filestore_op_threads 8
-/osd/filestore_max_sync_interval 5
-/osd/osd_max_scrubs 1
-
-#recovery tuning
-/osd/osd_recovery_max_active 5
-/osd/osd_max_backfills 2
-/osd/osd_recovery_op_priority 2
-/osd/osd_client_op_priority 63
-/osd/osd_recovery_max_chunk 1048576
-/osd/osd_recovery_threads 1
-
-#ports
-/osd/ms_bind_port_min 6800
-/osd/ms_bind_port_max 7100
-
-#client
-/client/rbd_cache_enabled true
-/client/rbd_cache_writethrough_until_flush true
-
-#mds
-/mds/mds_cache_size 100000
