--- ceph-releases/kraken/ubuntu/14.04/daemon/remove-mon.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-ceph mon remove $(hostname -s)
