--- ceph-releases/kraken/ubuntu/14.04/demo/s3cfg
@@ -1,44 +1,0 @@
-[default]
-access_key = AWS_ACCESS_KEY_PLACEHOLDER
-secret_key = AWS_SECRET_KEY_PLACEHOLDER
-bucket_location = US
-cloudfront_host = cloudfront.amazonaws.com
-default_mime_type = binary/octet-stream
-delete_removed = False
-dry_run = False
-enable_multipart = True
-encoding = ANSI_X3.4-1968
-encrypt = False
-follow_symlinks = False
-force = False
-get_continue = False
-gpg_command = /usr/bin/gpg
-gpg_decrypt = %(gpg_command)s -d --verbose --no-use-agent --batch --yes --passphrase-fd %(passphrase_fd)s -o %(output_file)s %(input_file)s
-gpg_encrypt = %(gpg_command)s -c --verbose --no-use-agent --batch --yes --passphrase-fd %(passphrase_fd)s -o %(output_file)s %(input_file)s
-gpg_passphrase =
-guess_mime_type = True
-host_base = localhost
-host_bucket = localhost
-human_readable_sizes = False
-invalidate_on_cf = False
-list_md5 = False
-log_target_prefix =
-mime_type =
-multipart_chunk_size_mb = 15
-preserve_attrs = True
-progress_meter = True
-proxy_host =
-proxy_port = 0
-recursive = False
-recv_chunk = 4096
-reduced_redundancy = False
-send_chunk = 4096
-simpledb_host = sdb.amazonaws.com
-skip_existing = False
-socket_timeout = 300
-urlencoding_mode = normal
-use_https = False
-verbosity = WARNING
-website_endpoint = http://%(bucket)s.s3-website-%(location)s.amazonaws.com/
-website_error =
-website_index = index.html
