+++ ceph-releases/kraken/ubuntu/16.04/daemon/2.sh
@@ -1,1 +1,6 @@
-../../14.04/daemon/2.sh
+#!/bin/sh
+
+PATH=/command:/usr/local/bin:/usr/local/sbin:/bin:/sbin:/usr/bin:/usr/sbin:/usr/X11R6/bin
+
+exec env - PATH=$PATH \
+runsvdir -P /etc/service 'log: ...........................................................................................................................................................................................................................................................................................................................................................................................................'
