--- ceph-releases/kraken/ubuntu/16.04/daemon/confd
@@ -1,1 +1,0 @@
-../../14.04/daemon/confd
