+++ ceph-releases/kraken/ubuntu/16.04/daemon/confd/conf.d/ceph.conf.toml.in
