+++ ceph-releases/kraken/ubuntu/16.04/daemon/confd/templates/ceph.conf.tmpl
