--- ceph-releases/kraken/ubuntu/16.04/daemon/my_init
@@ -1,1 +1,0 @@
-../../14.04/daemon/my_init
