+++ ceph-releases/kraken/ubuntu/16.04/daemon/my_init/my_init
