+++ ceph-releases/kraken/ubuntu/16.04/daemon/remove-mon.sh
@@ -1,1 +1,3 @@
-../../14.04/daemon/remove-mon.sh
+#!/bin/bash
+
+ceph mon remove $(hostname -s)
